Moved input_sample_files code to superclass
